move create command to archiver.create
