+++ setup.cfg
@@ -120,6 +120,7 @@ per_file_ignores =
     src/borg/archiver/benchmarks.py:F405
     src/borg/archiver/config.py:F405,E722
     src/borg/archiver/common.py:E501,F405
+    src/borg/archiver/create.py:E501,F405
     src/borg/archiver/debug.py:F405
     src/borg/archiver/delete.py:F405
     src/borg/archiver/diff.py:F405
